Refactor InfoCardTrait out of CardTrait
